requirements.txt
reflex==0.5.10
python-dotenv
supabase
